requirements.txt
blinker
certifi
charset-normalizer
click
DAWG-Python
dnspython
docopt
email_validator
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
idna
itsdangerous
Jinja2
lxml
MarkupSafe
pygame
pymorphy3
pymorphy3-dicts-ru
requests
SQLAlchemy
typing_extensions
urllib3
Werkzeug
WTForms
WTForms-SQLAlchemy
XlsxWriter